Move Token events to core_event_dispatcher
