+++ hook_event_dispatcher.module
@@ -5,7 +5,6 @@
  * Hook event dispatcher module.
  */
 
-use Drupal\Core\Render\BubbleableMetadata;
 use Drupal\Core\Session\AccountInterface;
 use Drupal\Core\Url;
 use Drupal\hook_event_dispatcher\Event\Cron\CronEvent;
@@ -13,8 +12,6 @@ use Drupal\hook_event_dispatcher\Event\Language\LanguageSwitchLinksAlterEvent;
 use Drupal\hook_event_dispatcher\Event\Path\PathDeleteEvent;
 use Drupal\hook_event_dispatcher\Event\Path\PathInsertEvent;
 use Drupal\hook_event_dispatcher\Event\Path\PathUpdateEvent;
-use Drupal\hook_event_dispatcher\Event\Token\TokensInfoEvent;
-use Drupal\hook_event_dispatcher\Event\Token\TokensReplacementEvent;
 use Drupal\hook_event_dispatcher\Event\Toolbar\ToolbarAlterEvent;
 use Drupal\hook_event_dispatcher\Event\User\UserCancelEvent;
 use Drupal\hook_event_dispatcher\Event\User\UserCancelMethodsAlterEvent;
@@ -48,41 +45,6 @@ function hook_event_dispatcher_help(string $routeName) {
     return '<pre>' . $output . '</pre>';
   }
   return '';
-}
-
-/**
- * Implements hook_tokens().
- *
- * {@inheritdoc}
- */
-function hook_event_dispatcher_tokens(
-  string $type,
-  array $tokens,
-  array $data,
-  array $options,
-  BubbleableMetadata $bubbleableMetadata
-) {
-  /* @var \Drupal\hook_event_dispatcher\Manager\HookEventDispatcherManagerInterface $manager */
-  $manager = Drupal::service('hook_event_dispatcher.manager');
-  $event = new TokensReplacementEvent($type, $tokens, $data, $options, $bubbleableMetadata);
-  $manager->register($event);
-  return $event->getReplacementValues();
-}
-
-/**
- * Implements hook_token_info().
- *
- * {@inheritdoc}
- */
-function hook_event_dispatcher_token_info() {
-  /* @var \Drupal\hook_event_dispatcher\Manager\HookEventDispatcherManagerInterface $manager */
-  $manager = Drupal::service('hook_event_dispatcher.manager');
-  $event = new TokensInfoEvent();
-  $manager->register($event);
-  return [
-    'types' => $event->getTokenTypes(),
-    'tokens' => $event->getTokens(),
-  ];
 }
 
 /**
