+++ modules/core_event_dispatcher/core_event_dispatcher.module
@@ -13,6 +13,7 @@ use Drupal\Core\Entity\EntityTypeInterface;
 use Drupal\Core\Field\FieldDefinitionInterface;
 use Drupal\Core\Field\FieldItemListInterface;
 use Drupal\Core\Form\FormStateInterface;
+use Drupal\Core\Render\BubbleableMetadata;
 use Drupal\Core\Session\AccountInterface;
 use Drupal\core_event_dispatcher\Event\Block\BlockBuildAlterEvent;
 use Drupal\core_event_dispatcher\Event\Entity\EntityAccessEvent;
@@ -51,6 +52,8 @@ use Drupal\core_event_dispatcher\Event\Theme\ThemeRegistryAlterEvent;
 use Drupal\core_event_dispatcher\Event\Theme\ThemesInstalledEvent;
 use Drupal\core_event_dispatcher\Event\Theme\ThemeSuggestionsAlterEvent;
 use Drupal\core_event_dispatcher\Event\Theme\ThemeSuggestionsAlterIdEvent;
+use Drupal\core_event_dispatcher\Event\Token\TokensInfoEvent;
+use Drupal\core_event_dispatcher\Event\Token\TokensReplacementEvent;
 
 // Block.
 // @todo hook_block_view_alter()
@@ -96,6 +99,10 @@ use Drupal\core_event_dispatcher\Event\Theme\ThemeSuggestionsAlterIdEvent;
 // @todo hook_js_settings_alter()
 // @todo hook_css_alter()
 // @todo hook_page_attachments_alter()
+
+// Token.
+// @todo hook_tokens_alter()
+// @todo hook_token_info_alter()
 
 /**
  * Implements hook_block_build_alter().
@@ -513,4 +520,39 @@ function core_event_dispatcher_template_preprocess_default_variables_alter(array
   /* @var \Drupal\hook_event_dispatcher\Manager\HookEventDispatcherManagerInterface $manager */
   $manager = Drupal::service('hook_event_dispatcher.manager');
   $manager->register(new TemplatePreprocessDefaultVariablesAlterEvent($variables));
+}
+
+/**
+ * Implements hook_tokens().
+ *
+ * {@inheritdoc}
+ */
+function core_event_dispatcher_tokens(
+  string $type,
+  array $tokens,
+  array $data,
+  array $options,
+  BubbleableMetadata $bubbleableMetadata
+) {
+  /* @var \Drupal\hook_event_dispatcher\Manager\HookEventDispatcherManagerInterface $manager */
+  $manager = Drupal::service('hook_event_dispatcher.manager');
+  $event = new TokensReplacementEvent($type, $tokens, $data, $options, $bubbleableMetadata);
+  $manager->register($event);
+  return $event->getReplacementValues();
+}
+
+/**
+ * Implements hook_token_info().
+ *
+ * {@inheritdoc}
+ */
+function core_event_dispatcher_token_info() {
+  /* @var \Drupal\hook_event_dispatcher\Manager\HookEventDispatcherManagerInterface $manager */
+  $manager = Drupal::service('hook_event_dispatcher.manager');
+  $event = new TokensInfoEvent();
+  $manager->register($event);
+  return [
+    'types' => $event->getTokenTypes(),
+    'tokens' => $event->getTokens(),
+  ];
 }
